Add TPC-DS benchmark support
Add Postgres role to build Postgres from a specific revision, along with
a git and systemd-user role to support this Postgres role.

Add tpcds role which will build the TPC-DS tools on the target host and
then generate queries and data on the target host before running the
queries there.

After the benchmark is run, the results are copied to a results
directory where an iPython notebook depicts the throughput for all
queries.

Also, README updates.