Add and attach managed disk
Explicitly add a managed storage account in the Azure VM provisioning.
Then use the Azure CLI directly to attach the disk. This is a hack to
get around a bug in the Ansible Azure module detailed in the code in
which the link of a managed data disk and a VM through the Ansible
Azure module is broken. This means you will have to skip the attach-disk
task explicitly every time you re-run the playbook after the first time.

Move all remote setup (done on the provisioned Azure VM) into a new set
of tasks, `remote.yaml`. This will format and partition the disk and
install the filesystem. It also will install the newer Linux kernel and
reboot the VM.