Make results location consistent
Make sub-directories in `results`, one for fio and one for tpcds.
Update the fetch commands in the tasks that copy the results over.
Also, update the iPython notebooks that have the source of the data
hard-coded.

Sadly, when I changed the path in the iPython notebook, I also checked
if it worked and thus some iPython notebook metadata changed making this
commit thoroughly annoying. There are apparently ways people deal with
this (versioning iPython notebooks in git):
https://pascalbugnion.net/blog/ipython-notebooks-and-git.html
However, I'm not sure this is the best solution long-term, so, do
nothing for now.