Use database for fio result run info
Instead of capturing run metadata in an index file, insert it into a run
table in a database.
The fio role uses an existing database with a specific schema. The user
must create the database using the schema specified in
roles/fio/files/schema.sql and provide connection details to the fio
role either by adding them to the fio role defaults or passing them on
the command line when running fio.yaml.

For compatability, schema.sql also creates a `run_staging` table which
is used by load.py to load existing fio result index files and their
associated run results into the database.

To use load.py, fill in the database connection information as well as
the location of the index and result files to be loaded.