Add workload concept
Specify fio job in a separate file and include workload id in run so
that run items can be compared only to other run items from the same
workload. I included workload_size so that I could still compare results
from running a workload with less data on a smaller disk to one with
more data on a larger disk (but with all the same other workload
parameters), in case that is needed.