Mount device instead of partition
Format entire drive instead of one partition and mount it. This works
around an issue with saving the partition UUID (not PARTUUID) for
mounting again later after unmounting.

Also, Azure seems to mount a drive at /mnt, so change our default mount
point to something completely different to be safe.